- revert the changes to the filter implementation in phoenicis-apps
