+++ phoenicis-apps/src/main/java/org/phoenicis/apps/Repository.java
@@ -21,7 +21,6 @@ package org.phoenicis.apps;
 import org.phoenicis.apps.dto.ApplicationDTO;
 import org.phoenicis.apps.dto.CategoryDTO;
 import org.phoenicis.apps.dto.ScriptDTO;
-import org.phoenicis.apps.filter.CombinedAppsFilter;
 
 import java.util.Collections;
 import java.util.List;
